add segformer, vanilla_unet eval and remove redundant demo_utils for imagenet
