+++ models/demos/segformer/demo/demo_for_image_classification.py
@@ -11,12 +11,12 @@ from ttnn.model_preprocessing import preprocess_model_parameters
 
 import ttnn
 from models.demos.segformer.common import load_config, load_torch_model
-from models.demos.segformer.demo.classification_demo_utils import get_batch, get_data_loader
 from models.demos.segformer.reference.segformer_for_image_classification import SegformerForImageClassificationReference
 from models.demos.segformer.tests.pcc.test_segformer_for_image_classification import create_custom_mesh_preprocessor
 from models.demos.segformer.tests.pcc.test_segformer_model import move_to_device
 from models.demos.segformer.tt.common import get_mesh_mappers
 from models.demos.segformer.tt.ttnn_segformer_for_image_classification import TtSegformerForImageClassification
+from models.demos.utils.common_demo_utils import get_batch, get_data_loader, load_imagenet_dataset
 
 
 def run_segformer_classification_demo(
@@ -24,7 +24,7 @@ def run_segformer_classification_demo(
 ):
     image_processor = AutoImageProcessor.from_pretrained("nvidia/mit-b0")
     logger.info("ImageNet-1k validation Dataset")
-    input_loc = str(model_location_generator("ImageNet_data"))
+    input_loc = load_imagenet_dataset(model_location_generator)
     correct = 0
     torch_correct = 0
     ttnn_correct = 0
@@ -33,9 +33,7 @@ def run_segformer_classification_demo(
     inputs_mapper, wts_mapper, output_composer = get_mesh_mappers(device)
 
     for iter in range(iterations):
-        inputs, labels = get_batch(data_loader)
-        inputs = image_processor(inputs, return_tensors="pt")
-        torch_input_tensor = inputs.pixel_values
+        torch_input_tensor, labels = get_batch(data_loader, image_processor)
         torch_input_tensor_permuted = torch.permute(torch_input_tensor, (0, 2, 3, 1))
         ttnn_input_tensor = ttnn.from_torch(
             torch_input_tensor_permuted,
